add auth, login

+++ backend/users/service/user_service.py
@@ -35,8 +35,33 @@ def get_user_by_id_services(id : int) -> User:
     cursor.execute("SELECT id, username, email FROM accounts WHERE id=%s", 
                     (str(id)))
     
+    if cursor.rowcount == 0:
+        return None
+    
     result = list(cursor.fetchone())
     result = User(result)
+
+    conn.close()
+    
+    return result
+
+def get_user_by_email_services(email: str) -> UserInDB:
+    conn = psycopg2.connect(
+        host=host_name,
+        port=port_number,
+        user=user_name,
+        password=password_db,
+        database=database_name
+    )
+    cursor = conn.cursor()
+    cursor.execute("SELECT * FROM accounts WHERE email=%s", 
+                    (email,))
+    
+    if cursor.rowcount == 0:
+        return None
+    
+    result = list(cursor.fetchone())
+    result = UserInDB(result)
 
     conn.close()
     
@@ -65,8 +90,16 @@ def update_user_service(id: int, username: str, email: str):
         database=database_name
     )
     cursor = conn.cursor()
+    
+    cursor.execute("SELECT id, username, email FROM accounts WHERE id=%s", 
+                    (str(id)))
+    
+    if cursor.rowcount == 0:
+        return None
+    
     cursor.execute("UPDATE accounts SET username=%s, email=%s WHERE id=%s", 
                     (username, email, id))
+    
     conn.commit()
     conn.close()
 
@@ -79,6 +112,13 @@ def delete_user_service(id: int):
         database=database_name
     )
     cursor = conn.cursor()
+
+    cursor.execute("SELECT id, username, email FROM accounts WHERE id=%s", 
+                    (str(id)))
+    
+    if cursor.rowcount == 0:
+        return None
+    
     cursor.execute("DELETE FROM accounts WHERE id=%s", (id,))
     conn.commit()
     conn.close()
